requirements_telegram.txt
# Только Telegram Bot API
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn[standard]==0.24.0
